AI_Trip_planner/AI_Trip_planner/requirements.txt
crewai
langchain-openai
serper-dev
langchain-huggingface
langchain-community
duckduckgo-search
huggingface_hub
python-dotenv